% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InoculationCalculation.R
\name{InoculationCalculation}
\alias{InoculationCalculation}
\title{Inoculum calculator}
\usage{
InoculationCalculation(
  stockTitre,
  inoculumTitre,
  inoculumVolume,
  replicates = 1,
  margin = 10
)
}
\arguments{
\item{stockTitre}{The titre/concentration of your stock}

\item{inoculumTitre}{The desired inoculum titre/concentration}

\item{inoculumVolume}{The amount of inoculum that is going to be added to a flask/well}

\item{replicates}{The total number of replicates that inoculum needs to be prepared for (default = 1)}

\item{margin}{The additional percentage of inoculum volume that should be prepared as a margin of error (default = 10)}
}
\value{
Returns a tibble listing the desired titre, the total volume of the initial stock required, the total volume of diluent required, and the total volume.
}
\description{
Calculate the required dilutions for a given inoculation concentration. If replicates are provided then the returned volumes are the total for all replicates summed.
}
\examples{
# Basic usage
InoculationCalculation(stockTitre = 10^7.8, inoculumTitre = 16000000, inoculumVolume = 300, replicates = 3, margin = 0)
#   `Inoculum Titre` `Stock Volume` `Diluent Volume` `Total volume`
#   <dbl>          <dbl>            <dbl>          <dbl>
#   1         16000000            228              672            900
}
