% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlasmidCopies.R
\name{PlasmidCopies}
\alias{PlasmidCopies}
\title{Plasmid copy calculator}
\usage{
PlasmidCopies(plasmidLength, insertLength, totalLength, weight)
}
\arguments{
\item{plasmidLength}{The length of the plasmid in base pairs}

\item{insertLength}{The length of your insert in base pairs}

\item{totalLength}{A combined total length of your plasmid and insert in base pairs, use this instead of plasmidLength and insertLength individually}

\item{weight}{The mass of your DNA as measured by spectrophotometer or other method, in nanograms}
}
\value{
a number
}
\description{
Plasmid copy calculator
}
\examples{
PlasmidCopies(totalLength = 3460, weight = 5.9)
1555867928

PlasmidCopies(plasmidLength = 3000, insertLength = 460, weight = 5.9)
1555867928

}
